feat: downloader supports multiple workers
