+++ src/downloader/request.ts
@@ -1,0 +1,23 @@
+import { IncomingMessage } from 'http';
+import { get } from 'https';
+
+export const request = (url: string): Promise<IncomingMessage> =>
+  new Promise((resolve) => {
+    const request = get(url, (response) => {
+      if (response.statusCode == 302) {
+        const redirectRequest = get(response.headers.location, resolve);
+        redirectRequest.on('error', (error) => {
+          // eslint-disable-next-line no-console
+          console.warn('request redirected error:', error.message);
+          throw error;
+        });
+      } else {
+        resolve(response);
+      }
+    });
+    request.on('error', (error) => {
+      // eslint-disable-next-line no-console
+      console.warn('request error:', error.message);
+      throw error;
+    });
+  });
